[ 1833671 ] [Launch] run ghci as shortcut key
The project executable and a GHCi session can now be launched as key shortcut. The behaviour is the same as from the Run As ... context menu. The key combinations are M3+M2 X H for the project executable and M3+M2 X G for GHCi. (M2 and M3 are the modificator keys, e.g. on Macs, the combination is Alt+Shift X H).

darcs-hash:20080329205621-34f1c-65e0770d62cf8b03979fa62af7376172f0ecdb32.gz