Improved the way GHC output is processed.
Objective: to get the GHC output in the Eclipse Console back.
Reinstated the ProcessRunner and ProcessFactory, but modified them a little.
Refactored parts of the stream processing code, though most is still similar to the old.
Moved the process runner stuff from ghccompiler to core, where it fits in better.
Should also port the QueryUtil to use these classes.