Added possibility to create project from existing source.
Just like in JDT, there's now a field in the New Project wizard where
a directory for the new project can be set. If this directory exists,
its contents are imported into the new project.

In fact, the new project is just created in the specified location
without regard for what's already there. More advanced importing
(e.g. detection of Cabal file, detection of source directories)
remains to be done.

The UI additions required some changes to the creation wizards,
leading to an overhaul of the Cabal import wizard as well.

In particular, there is now a common structure for wizard validators.
This is not as clean as I would like it to be, yet.