rollback_101010
Roll back out of my last series of changes because they were a bad idea.
What is needed: not a complete revamp of ScionManager and instance
factories, but executable factories. We want a null executable
(something that doesn't send or recieve messages), a built-in executable
and user-specified executable. When preferences change, the executable
factory changes and we update the current instances with new
executables. Less damage to what already exists.