scionmanager_101029
ScionPP dialog calls ScionManager directly, so that
ScionManager can figure out what should be done with the
new Scion/Cabal preferences. Given the choices between
use built-in, server flavor, etc., preference listeners are not
sufficient.

If we were doing this the Eclipse way, we'd generate an event.
That's probably complete overkill.

Make logging a little less verbose; we were logging messages
sent to the scion-server twice.