extdef_101109
External definitions for file URLs: Don't go through all of the extra
machinations of converting the URL to a URI, since that causes issuses
when the file path has spaces in it (URI won't parse). Just convert the
URL's path to a File and see if it exists.

Also, make the file URL in the preference page a proper file URL.