buildproject_101112
Refactor ScionInstance.buildProject() so that the user gets more
feedback via IProgressMonitor and, when projects get build via the
workspace build process, EclipseFP doesn't schedule additional
Jobs.

Also refactored the command group classes to separate out the
scheduling rules. It turns out that this was probably unnecessary
in hindsight, because the workspace build process already holds
the rule for the workspace and we just need to contribute a rule
for the project. Nonetheless, this does keep other EclipseFP
command groups from running because they're aware of rules
derived from IResource.