scionbuild_101112
Make the built-in scion server build process a little more annoying by
making the build Job a user job -- this pops up a nice little box that
tells the user that the scion server is building.

If successful, the build console is closed and removed. Could keep it
around, but what's the point?