Compressed outline view.
When a data declaration contains a single constructor with a name identical
to the data declaration, this hides the node for the constructor and
expands the constructor's children directly.  The intent is to reduce the
busy-ness of the outline view and the amount of repetative clicking to expand
it, without ever hiding any information.