Automate builds with Eclipse Tycho
While Eclipse Tycho is unbelievably awful, it's still better than
nothing. Apparently a good percentage (50%+) of Eclipse plug-ins are
using it for their builds.

To invoke:

* All plug-ins may be built using "mvn compile".
* All tests may be run using "mvn integration-test."

Currently, there are failing tests and they are ignored so that Maven
will continue to run the remaining tests. A number of these failures
are probably due to the test harness. Once the test failures are fixed,
we can disable ignoring test failures. After that, we can set up
continuous integration with Travis CI.

Other alterations:

* Added a root `.project` file so root files like the parent pom were
easier to edit in Eclipse.
* Added `target/` to root `.gitignore` since Maven uses `target/`
subdirectories for builds.

As an added bonus, easymock is now obtained from Eclipse Orbit. This
should mean that we can drop easymock from the repository.